+++ test/protractor/pages/home.js
@@ -1,9 +1,7 @@
 /*
- * Identity provider home page.
- *
- * Copyright (c) 2015 The Open Payments Foundation. All rights reserved.
+ * Copyright (c) 2015-2016 Digital Bazaar, Inc. All rights reserved.
  */
-var bedrock = GLOBAL.bedrock;
+var bedrock = global.bedrock;
 
 var api = {};
 module.exports = api;