+++ test/protractor/pages/idp.js
@@ -1,14 +1,12 @@
 /*
- * Identity provider page API.
- *
- * Copyright (c) 2015 The Open Payments Foundation. All rights reserved.
+ * Copyright (c) 2015-2016 Digital Bazaar, Inc. All rights reserved.
  */
-var bedrock = GLOBAL.bedrock;
-var browser = GLOBAL.browser;
-var by = GLOBAL.by;
-var element = GLOBAL.element;
-var expect = GLOBAL.expect;
-var should = GLOBAL.should;
+var bedrock = global.bedrock;
+var browser = global.browser;
+var by = global.by;
+var element = global.element;
+var expect = global.expect;
+var should = global.should;
 var api = {};
 module.exports = api;