+++ test/protractor/pages/index.js
@@ -1,7 +1,10 @@
-var pages = GLOBAL.bedrock.pages || {};
+/*
+ * Copyright (c) 2015-2016 Digital Bazaar, Inc. All rights reserved.
+ */
+var pages = global.bedrock.pages || {};
 
 pages.home = require('./home');
 pages.idp = require('./idp');
 pages.join = require('./join');
 
-module.exports = GLOBAL.bedrock.pages = pages;
+module.exports = global.bedrock.pages = pages;