+++ test/protractor/tests/e2e.js
@@ -1,56 +1,12 @@
-var bedrock = GLOBAL.bedrock;
-var expect = GLOBAL.expect;
-var should = GLOBAL.should;
-var describe = GLOBAL.describe;
-var it = GLOBAL.it;
+/*
+ * Copyright (c) 2015-2016 Digital Bazaar, Inc. All rights reserved.
+ */
+var bedrock = global.bedrock;
+var expect = global.expect;
+var should = global.should;
+var describe = global.describe;
+var it = global.it;
 var uuid = require('uuid');
-
-describe('IdP join form', function() {
-  var identity = {};
-  identity.sysIdentifier = bedrock.randomString().toLowerCase();
-  identity.id =
-    bedrock.baseUrl + bedrock.config.identityBasePath + identity.sysIdentifier;
-  identity.label = identity.sysIdentifier;
-  identity.email = identity.sysIdentifier + '@bedrock.dev';
-  identity.password = 'password';
-  var loginCredentials = {
-    sysIdentifier: identity.sysIdentifier,
-    passphrase: identity.password
-  };
-  var threeCharacters = 'abc';
-  var fortyCharacters = 'yA2NdBthMcnTqGYz3Eqe9uNHxM8u00TaooiuhIM';
-
-  beforeEach(function() {
-    bedrock.pages.join.get();
-  });
-
-  it('should contain the proper fields', function() {
-    bedrock.pages.join.checkFields();
-  });
-
-  it('should accept a valid form', function() {
-    bedrock.pages.join.createIdentity(identity);
-  });
-
-  // FIXME: write test for the various email related conditions
-  it('should warn on invalid email');
-
-  it('should warn on a short password', function() {
-    bedrock.pages.join.testField(
-      '$ctrl.identity.sysPassword', threeCharacters, 'minlength');
-  });
-
-  it('should warn on a long password', function() {
-    bedrock.pages.join.testField(
-      '$ctrl.identity.sysPassword', fortyCharacters, 'maxlength');
-  });
-
-  it('should warn if password confirmation does not match', function() {
-    bedrock.pages.join.testFieldsMatch(
-      '$ctrl.identity.sysPassword', '$ctrl.passphraseConfirmation',
-      'goodPhraseA', 'nonMatchingPhraseB', 'inputMatch');
-  });
-});
 
 /* FIXME: Below tests are currently broken */
 describe.skip('Navbar session management', function() {
@@ -175,4 +131,4 @@ function generateCredential() {
     }]
   };
   return mockCredential;
-};
+}